+++ models/wazuh-api.py
@@ -170,9 +170,11 @@ if __name__ == "__main__":
             'PUT/agents/:agent_id/restart': Agent.restart_agents,
             'PUT/agents/restart': Agent.restart_agents,
             'PUT/agents/:agent_name': Agent.add_agent,
+            'POST/agents/restart': Agent.restart_agents,
             'POST/agents': Agent.add_agent,
             'POST/agents/insert': Agent.insert_agent,
             'DELETE/agents/:agent_id': Agent.remove_agent,
+            'DELETE/agents/': Agent.remove_agent,
 
             '/agents/groups': Agent.get_all_groups,
             '/agents/groups/:group_id': Agent.get_agent_group,